+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix: differentiate between Identifier and TypeName nodes
 - Fix: don't allow illegal ref cycles
+- Fix: type names must start with uppercase character (or `_`)
 - Fix various other small bugs in the parser/checker
 - Massively improved unit test foundation
 - Don't retain comments as part of the AST (might be added back in later)